// arxflow.client/src/api/generated/model/createBoletaRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * ArxFlow.Server | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CreateBoletaRequest {
  /** @nullable */
  boletaPrincipalId?: number | null;
  /** @nullable */
  ativoId?: number | null;
  /** @nullable */
  contraparteId?: number | null;
  /** @nullable */
  fundoId?: number | null;
  /**
   * @maxLength 50
   * @nullable
   */
  ticker?: string | null;
  /** @maxLength 1 */
  tipoOperacao: string;
  volume?: number;
  quantidade?: number;
  tipoPrecificacao: string;
  /**
   * @maxLength 20
   * @nullable
   */
  ntnbReferencia?: string | null;
  /** @nullable */
  spreadValor?: number | null;
  /** @nullable */
  dataFixing?: string | null;
  /** @nullable */
  taxaNominal?: number | null;
  /** @nullable */
  pu?: number | null;
  /** @maxLength 100 */
  alocacao?: string;
  /** @maxLength 100 */
  usuario?: string;
  observacao?: string;
  /** @nullable */
  dataLiquidacao?: string | null;
  status?: string;
}